// frontend/src/app/api/models/config-inicio-panel.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ActorRef } from '../models/actor-ref';
import { ConfigInicioCampos } from '../models/config-inicio-campos';
import { ContenidoRefPanel } from '../models/contenido-ref-panel';
import { MedioMiniatura } from '../models/medio-miniatura';
export type ConfigInicioPanel = ConfigInicioCampos & {
'actualizado_en': string;
'actualizado_por': ActorRef;
'referencias': {
'medios'?: Array<MedioMiniatura>;
'contenidos'?: Array<ContenidoRefPanel>;
};
};