// frontend/src/app/api/models/cuenta.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { EstadoCuenta } from '../models/estado-cuenta';
import { Id } from '../models/id';
import { Rol } from '../models/rol';
export interface Cuenta {
  anonimizado_en?: (string | null);
  autorizacion_otorgada_en?: (string | null);
  autorizacion_version_politica?: (string | null);
  creado_en: string;
  debe_cambiar_credencial: boolean;
  desactivado_en?: (string | null);
  estado: EstadoCuenta;
  etiqueta: string;
  id: Id;
  mfa_activo: boolean;
  nombre_visible?: (string | null);
  rol: Rol;
  ultimo_acceso_en?: (string | null);

  /**
   * Null si ANONIMIZADA.
   */
  usuario?: (string | null);
}